fix(pacquet): keep approval prompts nonblocking
Preserve reporter output while a prompt owns the terminal and run the blocking dialog read
off the async worker. Prompt eligibility is explicit for deterministic install coverage, and all
fallible exclusion formatting now completes before the workspace manifest is written.

Constraint: Keep noninteractive installs fail-closed with the complete violation list.

Rejected: Drop prompt-time events | It loses logs and can strand the latest live frame.

Confidence: high

Scope-risk: narrow

Directive: Keep future interactive prompts inside pnpm:prompt start and end events.

Tested: release-age/reporter/install tests; affected-crate Clippy; rustfmt; diff checks.

Not-tested: Real TTY input.
